update the scope example
